fix main.c, correct libft.h and Makefile and add README

+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: fsoria <fsoria@student.42malaga.com>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/07 17:15:44 by fsoria            #+#    #+#             */
-/*   Updated: 2024/02/28 16:06:54 by fsoria           ###   ########.fr       */
+/*   Updated: 2024/02/28 19:40:34 by fsoria           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,8 +33,8 @@ void				*ft_memset(void *ptr, int x, int unsigned n);
 int					ft_strlen(const char *str);
 void				*ft_bzero(void *ptr, int unsigned n);
 void				*ft_memcpy(void *dest, const void *src, unsigned int n);
-void				*ft_memmove(void *dest, const void *src, int n);
-size_t				ft_strlcpy(char *dst, const char *src, int size);
+void				*ft_memmove(void *dest, const void *src, size_t n);
+size_t				ft_strlcpy(char *dst, const char *src, size_t size);
 size_t				ft_strlcat(char *dst, const char *src, size_t size);
 int					ft_toupper(int c);
 int					ft_tolower(int c);
